Refactored function to show dropdowns. Added an event to show header when in mobile.
